# Fix generation of inefficient MLD partitions