# Issue

See #6084 for details

# Performance

This has measurable improvement on MLD routing.
Will include test results in comments below.
TL;DR
- `/table` is ~6% faster
- `/route` is ~3% faster
- `osrm-routed` uses 0.2% less memory 

## Tasklist

 - [x] CHANGELOG.md entry ([How to write a changelog entry](http://keepachangelog.com/en/1.0.0/#how))
 - [ ] update relevant [Wiki pages](https://github.com/Project-OSRM/osrm-backend/wiki)
 - [ ] add tests (see [testing documentation](https://github.com/Project-OSRM/osrm-backend/blob/master/docs/testing.md)
 - [ ] review
 - [ ] adjust for comments
 - [ ] cherry pick to release branch

## Requirements / Relations

Closes #6084